Uploaded car.csv
I demonstrate object oriented programming in CSV using my car blueprint that I use in some of my other projects. The is only to store data, such as a C struct, with no member functions. To see an example of using this data in a project, visit my other repos, such as ObjectOrientedXMLWebDevelopment and ObjectOrientedJSONWebDevelopment.